+++ skorch/tests/test_helper.py
@@ -169,14 +169,79 @@ class TestSliceDict:
         print(gs.best_score_, gs.best_params_)
 
 
+class TestFilterParameterGroupsRequiresGrad():
+
+    @pytest.fixture
+    def filter_requires_grad(self):
+        from skorch.helper import filter_requires_grad
+        return filter_requires_grad
+
+    def test_all_parameters_requires_gradient(self, filter_requires_grad):
+        pgroups = [{
+            'params': [torch.zeros(1, requires_grad=True),
+                       torch.zeros(1, requires_grad=True)],
+            'lr': 0.1
+        }, {
+            'params': [torch.zeros(1, requires_grad=True)]
+        }]
+
+        filter_pgroups = list(filter_requires_grad(pgroups))
+        assert len(filter_pgroups) == 2
+        assert len(list(filter_pgroups[0]['params'])) == 2
+        assert len(list((filter_pgroups[1]['params']))) == 1
+
+        assert filter_pgroups[0]['lr'] == 0.1
+
+    def test_some_params_requires_gradient(self, filter_requires_grad):
+        pgroups = [{
+            'params': [
+                torch.zeros(1, requires_grad=True),
+                torch.zeros(1, requires_grad=False)
+            ], 'lr':0.1
+        }, {
+            'params': [torch.zeros(1, requires_grad=False)]
+        }]
+
+        filter_pgroups = list(filter_requires_grad(pgroups))
+        assert len(filter_pgroups) == 2
+        assert len(list(filter_pgroups[0]['params'])) == 1
+        assert len(list(filter_pgroups[1]['params'])) == 0
+
+        assert filter_pgroups[0]['lr'] == 0.1
+
+    def test_does_not_drop_group_when_requires_grad_is_false(
+            self, filter_requires_grad):
+        pgroups = [{
+            'params': [
+                torch.zeros(1, requires_grad=False),
+                torch.zeros(1, requires_grad=False)
+            ], 'lr':0.1
+        }, {
+            'params': [torch.zeros(1, requires_grad=False)]
+        }]
+
+        filter_pgroups = list(filter_requires_grad(pgroups))
+        assert len(filter_pgroups) == 2
+        assert len(list(filter_pgroups[0]['params'])) == 0
+        assert len(list(filter_pgroups[1]['params'])) == 0
+
+        assert filter_pgroups[0]['lr'] == 0.1
+
+
 class TestOptimizerParamsRequiresGrad:
 
     @pytest.fixture
-    def optimizer_params_requires_grad(self):
-        from skorch.helper import optimizer_params_requires_grad
-        return optimizer_params_requires_grad
+    def filtered_optimizer(self):
+        from skorch.helper import filtered_optimizer
+        return filtered_optimizer
 
-    def test_passes_filtered_cgroups(self, optimizer_params_requires_grad):
+    @pytest.fixture
+    def filter_requires_grad(self):
+        from skorch.helper import filter_requires_grad
+        return filter_requires_grad
+
+    def test_passes_filtered_cgroups(
+            self, filtered_optimizer, filter_requires_grad):
         pgroups = [{
             'params': [torch.zeros(1, requires_grad=True),
                        torch.zeros(1, requires_grad=False)],
@@ -185,8 +250,8 @@ class TestOptimizerParamsRequiresGrad:
             'params': [torch.zeros(1, requires_grad=True)]
         }]
 
-        filtered_func = optimizer_params_requires_grad(torch.optim.SGD)
-        filtered_opt = filtered_func(pgroups, lr=0.2)
+        opt = filtered_optimizer(torch.optim.SGD, filter_requires_grad)
+        filtered_opt = opt(pgroups, lr=0.2)
 
         assert isinstance(filtered_opt, torch.optim.SGD)
         assert len(list(filtered_opt.param_groups[0]['params'])) == 1
@@ -196,7 +261,7 @@ class TestOptimizerParamsRequiresGrad:
         assert filtered_opt.param_groups[1]['lr'] == 0.2
 
     def test_passes_kwargs_to_neuralnet_optimizer(
-            self, optimizer_params_requires_grad):
+            self, filtered_optimizer, filter_requires_grad):
         from skorch import NeuralNetClassifier
 
         class MyModule(torch.nn.Module):
@@ -207,10 +272,9 @@ class TestOptimizerParamsRequiresGrad:
             def forward(self, X):
                 return self.dense0(X)
 
+        opt = filtered_optimizer(torch.optim.SGD, filter_requires_grad)
         net = NeuralNetClassifier(
-            MyModule,
-            optimizer=optimizer_params_requires_grad(torch.optim.SGD),
-            optimizer__momentum=0.9)
+            MyModule, optimizer=opt, optimizer__momentum=0.9)
 
         net.initialize()
         assert isinstance(net.optimizer_, torch.optim.SGD)
